Updated by AS moving to directx 11(WIP) 1) fixed ScreenToWorld 2) removed warnings 3) vector2 is float instead of double
git-svn-id: https://sppong.googlecode.com/svn/trunk@265 c580c124-fe00-11dd-b556-9305e745af9e

+++ Source/Base/Includes/Color.inl
@@ -20,10 +20,10 @@ inline cColor::cColor( const int iRed, const int iGreen, const int iBlue, const 
 
 // *************************************************************************
 inline cColor::cColor(const float fRed, const float fGreen, const float fBlue, const float fAlpha)
-: m_iRed(fRed * 255)
-, m_iGreen(fGreen * 255)
-, m_iBlue(fBlue * 255)
-, m_iAlpha(fAlpha * 255)
+: m_iRed(static_cast<int>(fRed * 255))
+, m_iGreen(static_cast<int>(fGreen * 255))
+, m_iBlue(static_cast<int>(fBlue * 255))
+, m_iAlpha(static_cast<int>(fAlpha * 255))
 {
 	CalculateColorFromComponents();
 }
